task_queue.py modified

+++ LabelMatching/search/file_delete.py
@@ -3,13 +3,17 @@ import shutil
 
 def my_delete(target_addresses):          
     try:
-        for address in target_addresses:
-            if os.path.isfile(address):
-                os.remove(address)
-            elif os.path.isdir(address):
-                shutil.rmtree(address)
-            else:
-                pass
+        if len(target_addresses) == 0:
+            pass
+        else:
+            for address in target_addresses:
+                if os.path.isfile(address):
+                    os.remove(address)
+                elif os.path.isdir(address):
+                    shutil.rmtree(address)
+                else:
+                    pass
+        return 0
     except Exception as e:
         print(f"Error: {e}")
-        return None
+        return 2
